+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 
 - Add registered patients line-list report
 - Generic Vaccine line list report
+- Allow disabling reports through server config
 
 ## Tweaks ⚖️